Use Gmail-provided Unsubscribe Button if exists
Sometimes, Gmail will display a small, grey unsub button next to the
Sender info at the top of an email.  If that exists, click it, and then
click the blue Unsubscribe button that appears in a modal dialog
subsequently.

Fallback to looking for an unsubscribe button in the email body if this
method fails.

What is nice about trying to use the Gmail provided Unsubscribe buttons
first is that clicking them prevents you from having to open an new
window and click additional buttons on the Sender website.